+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.20.0 (coming soon)
+--------------------
+
+* !103: Revised the EnPT orthorectification to support the completely re-implemented sensormapgeo package (v1.0.0),
+  which leads to 30-50x faster orthorectification and adds support for all GDAL-supported resampling techniques.
+* !103: Fixed #58 (Holes in the orthorectified image in case of geographic target projection and bilinear resampling).
+
+
 0.19.7 (2024-01-04)
 -------------------